register in foreground service
