+++ app/src/main/java/io/appium/settings/ForegroundService.java
@@ -17,17 +17,38 @@
 package io.appium.settings;
 
 import android.app.Service;
+import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.os.IBinder;
 import android.util.Log;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
 import io.appium.settings.helpers.NotificationHelpers;
+import io.appium.settings.receivers.AnimationSettingReceiver;
+import io.appium.settings.receivers.BluetoothConnectionSettingReceiver;
+import io.appium.settings.receivers.ClipboardReceiver;
+import io.appium.settings.receivers.DataConnectionSettingReceiver;
+import io.appium.settings.receivers.HasAction;
+import io.appium.settings.receivers.LocaleSettingReceiver;
+import io.appium.settings.receivers.LocalesReader;
+import io.appium.settings.receivers.LocationInfoReceiver;
+import io.appium.settings.receivers.MediaScannerReceiver;
+import io.appium.settings.receivers.NotificationsReceiver;
+import io.appium.settings.receivers.SmsReader;
+import io.appium.settings.receivers.UnpairBluetoothDevicesReceiver;
+import io.appium.settings.receivers.WiFiConnectionSettingReceiver;
 
 public class ForegroundService extends Service {
     private static final String TAG = "APPIUM SERVICE";
     public static final String ACTION_START = "start";
     public static final String ACTION_STOP = "stop";
+
+    private final List<BroadcastReceiver> settingsReceivers = new ArrayList<>();
 
     @Override
     public IBinder onBind(Intent intent) {
@@ -49,6 +70,28 @@ public class ForegroundService extends Service {
         return super.onStartCommand(intent, flags, startId);
     }
 
+    @Override public void onCreate() {
+        registerSettingsReceivers(Arrays.asList(
+                WiFiConnectionSettingReceiver.class,
+                AnimationSettingReceiver.class,
+                DataConnectionSettingReceiver.class,
+                LocaleSettingReceiver.class,
+                LocalesReader.class,
+                LocationInfoReceiver.class,
+                ClipboardReceiver.class,
+                BluetoothConnectionSettingReceiver.class,
+                UnpairBluetoothDevicesReceiver.class,
+                NotificationsReceiver.class,
+                SmsReader.class,
+                MediaScannerReceiver.class
+        ));
+    }
+
+    @Override public void onDestroy() {
+        stopBroadCastReceiver();
+        super.onDestroy();
+    }
+
     private void startForegroundService() {
         startForeground(NotificationHelpers.APPIUM_NOTIFICATION_IDENTIFIER,
                 NotificationHelpers.getNotification(this));
@@ -66,5 +109,34 @@ public class ForegroundService extends Service {
         Intent intent = new Intent(context, ForegroundService.class);
         intent.setAction(ForegroundService.ACTION_START);
         return intent;
+    }
+
+    private void registerSettingsReceivers(List<Class<? extends BroadcastReceiver>> receiverClasses)
+    {
+        for (Class<? extends BroadcastReceiver> receiverClass: receiverClasses) {
+            try {
+                final BroadcastReceiver receiver = receiverClass.newInstance();
+                IntentFilter filter = new IntentFilter(((HasAction) receiver).getAction());
+                getApplicationContext().registerReceiver(receiver, filter);
+                Log.d(TAG, "Register " + receiver);
+                settingsReceivers.add(receiver);
+            } catch (IllegalAccessException e) {
+                Log.e(TAG, "Failed to register the receiver: " + receiverClass, e);
+            } catch (InstantiationException e) {
+                Log.e(TAG, "Failed to register the receiver: " + receiverClass, e);
+            }
+        }
+    }
+
+    private void stopBroadCastReceiver() {
+        for (BroadcastReceiver receiver: settingsReceivers) {
+            Log.d(TAG, "Unregister " + receiver);
+            try {
+                getApplicationContext().unregisterReceiver(receiver);
+            } catch (IllegalArgumentException e) {
+                // Can be ignored, so just for debugging purpose
+                Log.w(TAG, "Got an error in unregisterReceiver: " + receiver, e);
+            }
+        }
     }
 }
