+++ app/src/main/java/io/appium/settings/Settings.java
@@ -17,10 +17,8 @@
 package io.appium.settings;
 
 import android.app.Activity;
-import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
 import android.media.projection.MediaProjectionManager;
 import android.os.Build;
 import android.os.Bundle;
@@ -29,23 +27,6 @@ import android.util.Log;
 
 import java.io.File;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import io.appium.settings.receivers.AnimationSettingReceiver;
-import io.appium.settings.receivers.BluetoothConnectionSettingReceiver;
-import io.appium.settings.receivers.ClipboardReceiver;
-import io.appium.settings.receivers.DataConnectionSettingReceiver;
-import io.appium.settings.receivers.HasAction;
-import io.appium.settings.receivers.LocaleSettingReceiver;
-import io.appium.settings.receivers.LocalesReader;
-import io.appium.settings.receivers.LocationInfoReceiver;
-import io.appium.settings.receivers.MediaScannerReceiver;
-import io.appium.settings.receivers.NotificationsReceiver;
-import io.appium.settings.receivers.SmsReader;
-import io.appium.settings.receivers.UnpairBluetoothDevicesReceiver;
-import io.appium.settings.receivers.WiFiConnectionSettingReceiver;
 import io.appium.settings.recorder.RecorderService;
 import io.appium.settings.recorder.RecorderUtil;
 
@@ -74,7 +55,6 @@ public class Settings extends Activity {
     private int recordingMaxDuration = RECORDING_MAX_DURATION_DEFAULT_MS;
     private String recordingResolutionMode = NO_RESOLUTION_MODE_SET;
 
-    private final List<BroadcastReceiver> settingsReceivers = new ArrayList<>();
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -84,23 +64,8 @@ public class Settings extends Activity {
 
         String recordingAction = getIntent().getAction();
         if (recordingAction != null && recordingAction.startsWith(ACTION_RECORDING_BASE)) {
-            Log.d(TAG, "Skip unnecessary broadcast receiver registration for recording usage");
+            Log.d(TAG, "Skip starting foreground service");
         } else {
-            registerSettingsReceivers(Arrays.asList(
-                    WiFiConnectionSettingReceiver.class,
-                    AnimationSettingReceiver.class,
-                    DataConnectionSettingReceiver.class,
-                    LocaleSettingReceiver.class,
-                    LocalesReader.class,
-                    LocationInfoReceiver.class,
-                    ClipboardReceiver.class,
-                    BluetoothConnectionSettingReceiver.class,
-                    UnpairBluetoothDevicesReceiver.class,
-                    NotificationsReceiver.class,
-                    SmsReader.class,
-                    MediaScannerReceiver.class
-            ));
-
             // https://developer.android.com/about/versions/oreo/background-location-limits
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 startForegroundService(ForegroundService.getForegroundServiceIntent(Settings.this));
@@ -110,20 +75,6 @@ public class Settings extends Activity {
         }
 
         handleRecording(getIntent());
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        for (BroadcastReceiver receiver: settingsReceivers) {
-            try {
-                unregisterReceiver(receiver);
-            } catch (IllegalArgumentException e) {
-                // Can be ignored, so just for debugging purpose
-                Log.d(TAG, "Got an error in unregisterReceiver(" + receiver + "): " + e.getMessage());
-            }
-        }
-        ;
     }
 
     private void handleRecording(Intent intent) {
@@ -265,21 +216,5 @@ public class Settings extends Activity {
         startService(intent);
 
         finishActivity();
-    }
-
-    private void registerSettingsReceivers(List<Class<? extends BroadcastReceiver>> receiverClasses)
-    {
-        for (Class<? extends BroadcastReceiver> receiverClass: receiverClasses) {
-            try {
-                final BroadcastReceiver receiver = receiverClass.newInstance();
-                IntentFilter filter = new IntentFilter(((HasAction) receiver).getAction());
-                getApplicationContext().registerReceiver(receiver, filter);
-                settingsReceivers.add(receiver);
-            } catch (IllegalAccessException e) {
-                Log.e(TAG, "Failed to register the receiver: " + receiverClass, e);
-            } catch (InstantiationException e) {
-                Log.e(TAG, "Failed to register the receiver: " + receiverClass, e);
-            }
-        }
     }
 }
